LWSHADOOP-517: Prevent changes to CSVStrategy pseudo-constants
Some pieces of CSVIngestMapper fall back on predefined static
CSVStrategy objects, setup with reasonable defaults.  Other bits of
CSVIngestMapper modify the chosen strategy based on configuration
values.

The result is that the predefined CSVStrategy constants can drift from
their expected values over time, depending on how CSVIngestMapper is
used.  This was causing sporadic `hadoop-solr` test failures, based on
the order in which the tests were run.

This commit changes CSVIngestMapper to always use clones of the intended
strategy, so that the pre-defined constants actually stay constant.