LWSHADOOP-749: Test job-jar against Hadoop 2 and 3
Prior to this commit, we only tested this repo against Hadoop 2.  With
the release of now several versions in the Hadoop 3.x version line,
we should move towards testing and supporting Hadoop 3 as well.

This commit makes some gradle changes which extend the default 'test'
task, replacing it with two successors: `testHadoop2`, and
`testHadoop3`.  Both of these tasks run the same tests.  The only
different is the version of the Hadoop jars found on the classpaths when
running these tests.