LWSHADOOP-703: Use minicluster in CSVIngestMapperTest
Prior ot this commit, CSVIngestMapperTest used the 'mrunit' project for
its MR/Hadoop logic.  However, mrunit is a defunct project and will not
suppoprt recent versions of Hadoop, such as Hadoop 3.

This commit changes the first of several tests over to using an
in-memory Hadoop cluster (hadoop-minicluster) instead.  This allows us
to test the same logic, and is also more faithful to a production
environment than mrunit was.